requirements.txt
anyio==3.1.0
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.7.2
async-generator==1.10
attrs==21.2.0
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
backoff==1.11.1
bleach==3.3.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
entrypoints==0.3
filelock==3.12.0
idna==2.10
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.8.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==3.5.1
prometheus-client==0.11.0
prompt-toolkit==3.0.18
psycopg2==2.9.6
ptyprocess==0.7.0
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
ratelimit==2.2.1
requests==2.25.1
rise==5.7.1
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
Send2Trash==1.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.4
stevedore==5.1.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==4.6.3
urllib3==1.26.5
virtualenv==20.23.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
widgetsnbextension==3.5.1
xlrd==2.0.1
